multikernel: Track instance CPUs as 64-bit physical ID sets
Instance CPU assignments were NR_CPUS-sized bitmaps indexed by physical
CPU ID, which only works while physical IDs are small and dense. APIC
IDs on large x2APIC systems already exceed NR_CPUS, and the planned
arm64 and riscv ports use sparse 64-bit identifiers (MPIDR, hartid)
that cannot be bit positions at all.

Introduce mk_phys_cpu_t (u64) and struct mk_cpu_set, a small array with
add/del/contains/copy/format helpers. Entries keep insertion order, so
the first CPU listed in a device tree stays the instance's boot CPU,
matching the old find_first_bit() semantics for dense IDs. Returning
CPUs walks the request back-to-front so teardown can pass the
instance's own set without the walk skipping entries as they are
deleted.

Convert everything that indexed bitmaps by physical ID: instance and
config structures, transfer/return, kexec spawn CPU selection, KHO
restore, baseline offlining, hotplug messages, and the x86 spawn repark
loops. The arch helpers arch_cpu_physical_id() and
arch_cpu_from_physical_id() now use u64, as does mk_force_stop_cpu().

The formats that carry physical IDs widen with the model:

 - The "cpus" property in baseline and instance DTBs is now an array
   of 64-bit cells, on both the parse and generate sides, so the
   host-to-spawn KHO handoff stays consistent. Baselines written with
   32-bit cells are rejected with a clear error.
 - Overlay cpu@N nodes accept a reg of one 64-bit or one 32-bit cell,
   so existing overlays keep working.
 - mk_ipi_data.sender_cpu and mk_cpu_resource_payload.cpu_id are u64,
   and mk_resource_ack.resource_id widens to u64, which also stops
   memory ACKs from matching on a truncated (u32)start_pfn.

Also fixed while converting the surrounding code: sending to an
instance with no CPUs now fails with -ENODEV before claiming a ring
slot instead of raising an IPI at a garbage APIC ID computed from an
empty bitmap; the KHO restore error path no longer frees dtb_data
twice; and mk_dt_config_free() clears the CPU set pointer so the
parse-error path cannot double free it.

+++ arch/x86/include/asm/cpu.h
@@ -68,7 +68,7 @@ extern struct cpumask cpus_stop_mask;
 
 #ifdef CONFIG_MULTIKERNEL
 int mk_register_stop_nmi_handler(void);
-void mk_force_stop_cpu(int phys_cpu);
+void mk_force_stop_cpu(u64 phys_cpu);
 void __noreturn mk_enter_pool_state(void *info);
 #endif
